Add files via upload
Deep Learning with Python is a practical guide to understanding and implementing neural networks. Covering key architectures like CNNs and RNNs, it features hands-on examples using TensorFlow and Keras. Ideal for all skill levels, this resource equips readers to build, train, and deploy deep learning models for applications in computer vision and natural language processing. Unlock the power of deep learning in AI!